requirements.txt
httpx>=0.24.0
tenacity>=8.2.0
click>=8.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=7.3.0